Invalidate shader cache on GL driver mismatch
If I run daemon with Intel integrated graphics and then run it with
Nvidia, it will incorrectly try to load compiled shaders from one card
into the other. So store a hash of the GL vendor/version/etc. strings to
detect and prevent this situation.

When one such mismatch is detected, ignore the cache for all subsequent
shaders because this string is expected to be the same for all shaders.