Refactor NET_Config; drop cvar modified checks
NET_Config(bool enableNetworking) was a function to either open the
network sockets if the arg is true, or close them if it is false.
It checked the modified flag of a bunch of net_ cvars, with the idea
that if it is requested to enable networking while it was already
enabled, to not do anything unless the configuration is changed.
But actually this situation can never happen, so we can drop the modify
checks and always reinitialize.

NET_Config( true ) was only done in 4 places, none of which could
possibly omit the network (re-)initialization:
- NET_Init - only called once on startup
- NET_Restart - the point of /net_restart is to force reinitialization
- SV_Startup - this is only done in clients and starting the server is a
  state change for clients so restart can't be skipped
- SV_Shutdown - this is only done in clients and stopping the server is
  a state change for clients so restart can't be skipped.

Also split NET_Config into two functions NET_EnableNetworking and
NET_DisableNetworking for better readability.